+++ CHANGES.rst
@@ -4,6 +4,7 @@ CHANGELOG
 unreleased
 -------
 
+- [enhancements] command line, pytest.ini and fixture factory options for modifying connection timeout
 - [enhancements] command line and pytest.ini options for modifying executable
 - [enhancements] command line and pytest.ini options for modifying host
 - [enhancements] command line and pytest.ini options for modifying port